Backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# VoiceCraft Platform — requirements.txt
# All packages: free, open-source, available as of 2025
# ═══════════════════════════════════════════════════════════════

# Build tools — must be first
setuptools<70.0
wheel>=0.45

# ── Web Framework ──────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
websockets==13.1
httpx==0.28.0

# ── Async / Task Queue ─────────────────────────────────────────
celery==5.4.0
redis==5.2.0
kombu==5.4.2

# ── Database ───────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# ── Validation / Settings ─────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ── Auth / Security ────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# ── Object Storage (MinIO / S3-compatible, self-hosted FREE) ──
boto3==1.35.79
minio==7.2.11

# ── Core Audio Processing ─────────────────────────────────────
librosa==0.10.2
soundfile==0.12.1
numpy>=1.23.0,<1.25.0
scipy>=1.10.0,<1.13.0
pydub==0.25.1
audioread==3.0.1
resampy==0.4.3
noisereduce==3.0.3
webrtcvad-wheels==2.0.10.post2

# ── Voice Cloning & TTS (Coqui XTTS-v2 — Apache 2.0, FREE) ──
TTS==0.22.0
torch==2.4.1
torchaudio==2.4.1

# ── Speaker Verification / Embedding ─────────────────────────
speechbrain==1.0.2

# ── Deepfake Detection Models ─────────────────────────────────
# AASIST, RawNet2 — implemented locally, weights from HuggingFace (free)
transformers==4.46.3
huggingface-hub==0.26.2
accelerate==1.1.1
einops==0.8.0

# ── Audio Feature Extraction ──────────────────────────────────
python-speech-features==0.6
antropy==0.1.9          # Entropy features for liveness detection
praat-parselmouth==0.4.4 # Glottal / prosodic analysis

# ── ML / Inference ─────────────────────────────────────────────
scikit-learn==1.5.2
joblib==1.4.2
onnxruntime==1.20.1     # Fast CPU inference for detection models
pandas<2.0.0,>=1.4.0

# ── Transcription (Whisper — MIT License, FREE) ───────────────
openai-whisper

# ── Background Tasks / Concurrency ───────────────────────────
anyio==4.7.0
tenacity==9.0.0

# ── Monitoring / Observability ────────────────────────────────
prometheus-fastapi-instrumentator==7.0.0
structlog==24.4.0

# ── File Format Support ───────────────────────────────────────
ffmpeg-python==0.2.0
mutagen==1.47.0

# ── Utilities ─────────────────────────────────────────────────
python-slugify==8.0.4
shortuuid==1.0.13
aiofiles==24.1.0
cachetools==5.5.0
tqdm==4.67.1
psutil==6.1.0
